Produce a data frame to summarise adverse event data.

Takes a data frame of adverse events and produces the number of events and number and percentage of individuals with an adverse event.

Usage

ae_table(df, ..., group, ID, N, accuracy = 0.1, total = FALSE)

Arguments

df

Data Frame

...

Variables to be summarised

group

Variable that defines the grouping

ID

Variable that defines the individual identifier (e.g. screening number)

N

a data frame with the group counts (typically produced using dplyr::count)

accuracy

see details of scales::percent

total

Logical indicating whether a total column should be created